feat(db): extract db config, add sequelize-cli
- Move database connection config to src/config/database.ts
- Install sequelize-cli for migration management
- Create new folder structure under src/infra/database
- Add models, migrations, seeders, and sequelize.ts
