+++ .sequelizerc
@@ -1,0 +1,8 @@
+const path = require('path')
+
+module.exports = {
+  config: path.resolve('src', 'config', 'database.ts'),
+  'models-path': path.resolve('src', 'infra', 'database', 'models'),
+  'seeders-path': path.resolve('src', 'infra', 'database', 'seeders'),
+  'migrations-path': path.resolve('src', 'infra', 'database', 'migrations')
+}
