+++ src/config/database.ts
@@ -1,0 +1,18 @@
+import path from 'node:path'
+
+import { SequelizeOptions } from 'sequelize-typescript'
+
+import { env } from '@/config/env'
+
+const config: SequelizeOptions = {
+  dialect: 'postgres',
+  host: env.DATABASE_HOST,
+  port: env.DATABASE_PORT,
+  database: env.DATABASE_NAME,
+  username: env.DATABASE_USERNAME,
+  password: env.DATABASE_PASSWORD,
+  logging: env.NODE_ENV === 'development',
+  models: [path.resolve('..', 'models', '**', '*.ts')],
+}
+
+export default config
