--- src/env.ts
@@ -1,25 +1,0 @@
-import 'dotenv/config'
-
-import { z } from 'zod'
-
-const envSchema = z.object({
-  NODE_ENV: z
-    .enum(['development', 'test', 'production'])
-    .default('development'),
-  SERVER_PORT: z.coerce.number().default(3333),
-  DATABASE_HOST: z.string(),
-  DATABASE_PORT: z.coerce.number(),
-  DATABASE_NAME: z.string(),
-  DATABASE_USERNAME: z.string(),
-  DATABASE_PASSWORD: z.string(),
-})
-
-const _env = envSchema.safeParse(process.env)
-
-if (_env.success === false) {
-  console.error('Invalid environment variables', _env.error.format())
-
-  throw new Error('Invalid environment variables.')
-}
-
-export const env = _env.data
